+++ CHANGELOG.md
@@ -22,6 +22,10 @@
 - Updated LitEmoji ot 2.x.
 - Updated webonyx/graphql-php to 14.x.
 
+### Fixed
+- Fixed a PHP error that could occur on the System Report utility if Craft was installed using Composer 1.
+- Fixed a PHP error that could occur on the System Report utility if the wrong `Composer\Semver\VersionParser\InstalledVersions` class was autoloaded.
+
 ## 3.6.0-beta.2 - 2020-11-04
 
 ### Added